✨ 6CRM12 release: prepare v0.4.0 candidate
