+++ .agentplane/tasks/202604300956-6CRM12/pr/github-title.txt
@@ -1,0 +1,1 @@
+publish/release: Publish v0.4.0 release (6CRM12)
