+++ .agentplane/tasks/202604300956-6CRM12/pr/meta.json
@@ -1,0 +1,14 @@
+{
+  "base": "main",
+  "branch": "task/202604300956-6CRM12/v0-4-0-release",
+  "created_at": "2026-04-30T10:07:30.508Z",
+  "head_sha": "05ae6e59483b05524d3570e09f8c933f4581a7c6",
+  "last_verified_at": null,
+  "last_verified_sha": null,
+  "schema_version": 1,
+  "task_id": "202604300956-6CRM12",
+  "updated_at": "2026-04-30T10:07:30.508Z",
+  "verify": {
+    "status": "skipped"
+  }
+}
