+++ packages/agentplane/src/cli/run-cli.core.hooks.test.ts
@@ -1186,146 +1186,4 @@ describe("runCli", { timeout: HOOKS_SUITE_TIMEOUT_MS }, () => {
     expect(output).toContain("Skipping pre-push checks for delete-only remote branch cleanup.");
     expect(output).not.toContain("Running pre-push checks");
   });
-
-  it("hooks run pre-commit allows tasks.json with env override", async () => {
-    const root = await mkGitRepoRoot();
-    await writeDefaultConfig(root);
-    await writeFile(path.join(root, ".agentplane", "tasks.json"), "{}", "utf8");
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", ".agentplane/tasks.json"], { cwd: root });
-
-    const prev = process.env.AGENTPLANE_ALLOW_TASKS;
-    process.env.AGENTPLANE_ALLOW_TASKS = "1";
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli(["hooks", "run", "pre-commit", "--root", root]);
-      expect(code).toBe(0);
-    } finally {
-      io.restore();
-      if (prev === undefined) delete process.env.AGENTPLANE_ALLOW_TASKS;
-      else process.env.AGENTPLANE_ALLOW_TASKS = prev;
-    }
-  });
-
-  it("hooks run pre-commit blocks AGENTS.md without env override", async () => {
-    const root = await mkGitRepoRoot();
-    await writeDefaultConfig(root);
-    await writeFile(path.join(root, "AGENTS.md"), "# policy\n", "utf8");
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", "AGENTS.md"], { cwd: root });
-
-    const prev = process.env.AGENTPLANE_ALLOW_POLICY;
-    process.env.AGENTPLANE_ALLOW_POLICY = "0";
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli(["hooks", "run", "pre-commit", "--root", root]);
-      expect(code).toBe(5);
-      expect(io.stderr).toContain("AGENTS.md is protected by agentplane hooks");
-      expect(io.stderr).toContain("AGENTPLANE_ALLOW_POLICY=1");
-    } finally {
-      io.restore();
-      if (prev === undefined) delete process.env.AGENTPLANE_ALLOW_POLICY;
-      else process.env.AGENTPLANE_ALLOW_POLICY = prev;
-    }
-  });
-
-  it("hooks run pre-commit allows AGENTS.md with env override", async () => {
-    const root = await mkGitRepoRoot();
-    await writeDefaultConfig(root);
-    await writeFile(path.join(root, "AGENTS.md"), "# policy\n", "utf8");
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", "AGENTS.md"], { cwd: root });
-
-    const prev = process.env.AGENTPLANE_ALLOW_POLICY;
-    process.env.AGENTPLANE_ALLOW_POLICY = "1";
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli(["hooks", "run", "pre-commit", "--root", root]);
-      expect(code).toBe(0);
-    } finally {
-      io.restore();
-      if (prev === undefined) delete process.env.AGENTPLANE_ALLOW_POLICY;
-      else process.env.AGENTPLANE_ALLOW_POLICY = prev;
-    }
-  });
-
-  it("hooks run pre-commit allows base branch with allowBase", async () => {
-    const root = await mkGitRepoRootWithBranch("main");
-    const config = defaultConfig();
-    config.workflow_mode = "branch_pr";
-    await writeConfig(root, config);
-    await runCliSilent(["branch", "base", "set", "main", "--root", root]);
-    await mkdir(path.join(root, "src"), { recursive: true });
-    await writeFile(path.join(root, "src", "app.ts"), "x", "utf8");
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", "src/app.ts"], { cwd: root });
-
-    const prev = process.env.AGENTPLANE_ALLOW_BASE;
-    process.env.AGENTPLANE_ALLOW_BASE = "1";
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli(["hooks", "run", "pre-commit", "--root", root]);
-      expect(code).toBe(0);
-    } finally {
-      io.restore();
-      if (prev === undefined) delete process.env.AGENTPLANE_ALLOW_BASE;
-      else process.env.AGENTPLANE_ALLOW_BASE = prev;
-    }
-  });
-
-  it("hooks run pre-commit enforces branch_pr base restrictions", async () => {
-    const root = await mkGitRepoRootWithBranch("main");
-    const config = defaultConfig();
-    config.workflow_mode = "branch_pr";
-    await writeConfig(root, config);
-    await runCliSilent(["branch", "base", "set", "main", "--root", root]);
-    await mkdir(path.join(root, "src"), { recursive: true });
-    await writeFile(path.join(root, "src", "app.ts"), "x", "utf8");
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", "src/app.ts"], { cwd: root });
-
-    const prev = process.env.AGENTPLANE_ALLOW_BASE;
-    delete process.env.AGENTPLANE_ALLOW_BASE;
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli(["hooks", "run", "pre-commit", "--root", root]);
-      expect(code).toBe(5);
-      expect(io.stderr).toContain("forbidden on main");
-    } finally {
-      io.restore();
-      if (prev === undefined) delete process.env.AGENTPLANE_ALLOW_BASE;
-      else process.env.AGENTPLANE_ALLOW_BASE = prev;
-    }
-  });
-
-  it("hooks run pre-commit blocks tasks.json off base in branch_pr", async () => {
-    const root = await mkGitRepoRootWithBranch("main");
-    const config = defaultConfig();
-    config.workflow_mode = "branch_pr";
-    await writeConfig(root, config);
-    await runCliSilent(["branch", "base", "set", "main", "--root", root]);
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["checkout", "-b", "feature"], { cwd: root });
-    await writeFile(path.join(root, ".agentplane", "tasks.json"), "{}", "utf8");
-    await execFileAsync("git", ["add", ".agentplane/tasks.json"], { cwd: root });
-
-    const prev = process.env.AGENTPLANE_ALLOW_TASKS;
-    process.env.AGENTPLANE_ALLOW_TASKS = "1";
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli(["hooks", "run", "pre-commit", "--root", root]);
-      expect(code).toBe(5);
-      expect(io.stderr).toContain("allowed only on main");
-    } finally {
-      io.restore();
-      if (prev === undefined) delete process.env.AGENTPLANE_ALLOW_TASKS;
-      else process.env.AGENTPLANE_ALLOW_TASKS = prev;
-    }
-  });
 });
