+++ packages/agentplane/src/runtime/prompt-modules/compiler.ts
@@ -2,11 +2,17 @@ import type {
   PromptModule,
   PromptModuleAddress,
   PromptModuleGraph,
-  PromptModuleGraphNode,
   PromptModuleLoadCondition,
   PromptModuleMutability,
 } from "./model.js";
 import { PROMPT_MODULE_CONTRACT_SCHEMA_VERSION } from "./model.js";
+import { mergeDuplicateNodes } from "./compiler.merge.js";
+import {
+  copyModule,
+  moduleAddress,
+  uniqueStrings,
+  type WorkingPromptModuleNode,
+} from "./compiler.shared.js";
 import type {
   PromptModuleBinding,
   PromptModuleMutation,
@@ -64,15 +70,6 @@ export type PromptModuleCompiledGraph = PromptModuleGraph & {
   ok: boolean;
 };
 
-type WorkingPromptModuleNode = PromptModuleGraphNode & {
-  sequence: number;
-  disabled?: boolean;
-};
-
-function uniqueStrings(values: Iterable<string>): string[] {
-  return [...new Set([...values].filter((value) => value.trim().length > 0))];
-}
-
 function contextCommands(context: PromptModuleCompilerContext): string[] {
   return uniqueStrings([context.command ?? "", ...(context.commands ?? [])]);
 }
@@ -105,10 +102,6 @@ export function matchesPromptModuleLoadCondition(
     intersects(condition.repo_types, contextRepoTypes(context)) &&
     intersects(condition.recipe_ids, context.recipe_ids ?? [])
   );
-}
-
-function moduleAddress(module: PromptModule): string {
-  return module.address.value;
 }
 
 function moduleRecipeId(module: PromptModule): string | undefined {
@@ -176,10 +169,6 @@ function findMatchingNodeIndexes(
   return indexes;
 }
 
-function copyModule(module: PromptModule): PromptModule {
-  return structuredClone(module);
-}
-
 function patchModule(module: PromptModule, patch: PromptModuleStructuredPatch): PromptModule {
   const next = copyModule(module);
   if (patch.title !== undefined) next.title = patch.title;
@@ -229,201 +218,6 @@ function reportMutabilityViolation(
     mutation_id: mutation.id,
     module_address: moduleAddress(module),
     message: `${mutation.op} mutation ${mutation.id} cannot modify prompt module ${moduleAddress(module)} because its mutability is ${module.mutability}; only replaceable modules allow direct patch, replace, or disable mutations.`,
-  });
-}
-
-function precedenceOf(module: PromptModule): number {
-  return module.merge.precedence ?? 0;
-}
-
-function sortedByPrecedence(nodes: readonly WorkingPromptModuleNode[]): WorkingPromptModuleNode[] {
-  return nodes.toSorted((left, right) => {
-    const precedenceDiff = precedenceOf(left.module) - precedenceOf(right.module);
-    if (precedenceDiff !== 0) return precedenceDiff;
-    return left.sequence - right.sequence;
-  });
-}
-
-function chooseHighestPrecedence(
-  nodes: readonly WorkingPromptModuleNode[],
-): WorkingPromptModuleNode {
-  return sortedByPrecedence(nodes).at(-1)!;
-}
-
-function chooseLastWriter(nodes: readonly WorkingPromptModuleNode[]): WorkingPromptModuleNode {
-  return nodes.toSorted((left, right) => left.sequence - right.sequence).at(-1)!;
-}
-
-function cloneNode(node: WorkingPromptModuleNode): WorkingPromptModuleNode {
-  return {
-    ...node,
-    module: copyModule(node.module),
-    replaces: node.replaces ? [...node.replaces] : undefined,
-    extends: node.extends ? [...node.extends] : undefined,
-  };
-}
-
-function mergeStringContent(
-  nodes: readonly WorkingPromptModuleNode[],
-  direction: "append" | "prepend",
-): string | null {
-  if (nodes.some((node) => typeof node.module.content !== "string")) return null;
-  const sorted = sortedByPrecedence(nodes);
-  const ordered = direction === "append" ? sorted : sorted.toReversed();
-  return ordered
-    .map((node) => (typeof node.module.content === "string" ? node.module.content : ""))
-    .join("");
-}
-
-function mergeObjectContent(
-  nodes: readonly WorkingPromptModuleNode[],
-): Record<string, unknown> | null {
-  const merged: Record<string, unknown> = {};
-  for (const node of sortedByPrecedence(nodes)) {
-    if (!node.module.content || typeof node.module.content !== "object") return null;
-    if (Array.isArray(node.module.content)) return null;
-    Object.assign(merged, node.module.content);
-  }
-  return merged;
-}
-
-function unionByIdContent(
-  nodes: readonly WorkingPromptModuleNode[],
-): Record<string, unknown> | null {
-  const merged: Record<string, unknown> = {};
-  for (const node of sortedByPrecedence(nodes)) {
-    if (!node.module.content || typeof node.module.content !== "object") return null;
-    if (Array.isArray(node.module.content)) return null;
-    for (const [key, value] of Object.entries(node.module.content)) {
-      if (!Array.isArray(value)) {
-        merged[key] = value;
-        continue;
-      }
-      const byId = new Map<string, unknown>();
-      const previous = merged[key];
-      if (Array.isArray(previous)) {
-        for (const item of previous) {
-          const id = item && typeof item === "object" ? (item as { id?: unknown }).id : null;
-          if (typeof id === "string") byId.set(id, item);
-        }
-      }
-      for (const item of value) {
-        const id = item && typeof item === "object" ? (item as { id?: unknown }).id : null;
-        if (typeof id === "string") byId.set(id, item);
-      }
-      merged[key] = [...byId.values()];
-    }
-  }
-  return merged;
-}
-
-function mergeDuplicateNodes(
-  nodes: readonly WorkingPromptModuleNode[],
-  diagnostics: PromptModuleDiagnostic[],
-): WorkingPromptModuleNode[] {
-  const byAddress = new Map<string, WorkingPromptModuleNode[]>();
-  for (const node of nodes) {
-    const address = moduleAddress(node.module);
-    byAddress.set(address, [...(byAddress.get(address) ?? []), node]);
-  }
-
-  const merged: WorkingPromptModuleNode[] = [];
-  for (const [address, group] of byAddress.entries()) {
-    if (group.length === 1) {
-      const only = group[0];
-      if (only) merged.push(cloneNode(only));
-      continue;
-    }
-
-    const latest = group.at(-1);
-    const first = group[0];
-    if (!latest || !first) continue;
-    const conflictPolicy = latest.module.merge.conflict;
-    const mergeMode = latest.module.merge.mode;
-    switch (conflictPolicy) {
-      case "error": {
-        diagnostics.push({
-          severity: "error",
-          code: "duplicate_module",
-          module_address: address,
-          message: `Multiple active prompt modules target ${address}.`,
-        });
-        merged.push(cloneNode(first));
-        continue;
-      }
-      case "highest_precedence": {
-        merged.push(cloneNode(chooseHighestPrecedence(group)));
-        continue;
-      }
-      case "last_writer_wins": {
-        merged.push(cloneNode(chooseLastWriter(group)));
-        continue;
-      }
-      case "keep_all": {
-        if (mergeMode !== "replace") break;
-        merged.push(cloneNode(chooseLastWriter(group)));
-        continue;
-      }
-    }
-
-    const base = cloneNode(chooseHighestPrecedence(group));
-    switch (mergeMode) {
-      case "append":
-      case "prepend": {
-        const content = mergeStringContent(group, mergeMode);
-        if (content === null) {
-          diagnostics.push({
-            severity: "error",
-            code: "duplicate_module",
-            module_address: address,
-            message: `Cannot ${mergeMode} non-string prompt module content for ${address}.`,
-          });
-        } else {
-          base.module.content = content;
-        }
-        break;
-      }
-      case "merge_object": {
-        const content = mergeObjectContent(group);
-        if (content === null) {
-          diagnostics.push({
-            severity: "error",
-            code: "duplicate_module",
-            module_address: address,
-            message: `Cannot merge non-object prompt module content for ${address}.`,
-          });
-        } else {
-          base.module.content = content;
-        }
-        break;
-      }
-      case "union_by_id": {
-        const content = unionByIdContent(group);
-        if (content === null) {
-          diagnostics.push({
-            severity: "error",
-            code: "duplicate_module",
-            module_address: address,
-            message: `Cannot union non-object prompt module content for ${address}.`,
-          });
-        } else {
-          base.module.content = content;
-        }
-        break;
-      }
-      case "pick_one": {
-        break;
-      }
-    }
-    base.extends = uniqueStrings(group.flatMap((node) => node.extends ?? []));
-    base.replaces = uniqueStrings(group.flatMap((node) => node.replaces ?? []));
-    merged.push(base);
-  }
-
-  return merged.toSorted((left, right) => {
-    const precedenceDiff = precedenceOf(left.module) - precedenceOf(right.module);
-    if (precedenceDiff !== 0) return precedenceDiff;
-    return left.sequence - right.sequence;
   });
 }
